GoogleFunctions/function-source/requirements.txt
pinecone-client==2.2.1
sentence_transformers==2.2.2
Flask==2.3.2
Flask-Cors==3.0.10